+++ app/Http/Controllers/OutletController.php
@@ -85,26 +85,33 @@ class OutletController extends Controller
             ], 403);
         }
         $outlet = Outlet::find($id);
-        $filename = $outlet->image;
-        if ($file = $request->file('image')) {
-            $dir = 'uploads/outlet';
-            if (file_exists(public_path($dir.$filename))) {
-                unlink(public_path($dir.$filename));
+        if ($outlet) {
+            $filename = $outlet->image;
+            if ($file = $request->file('image')) {
+                $dir = 'uploads/outlet';
+                if (file_exists(public_path($dir.$filename))) {
+                    unlink(public_path($dir.$filename));
+                }
+                $filename = time().rand(1111,9999).'.'.$file->getClientOriginalExtension();
+                $file->move($dir, $filename);
             }
-            $filename = time().rand(1111,9999).'.'.$file->getClientOriginalExtension();
-            $file->move($dir, $filename);
+            $outlet->update([
+                "name" => $request->name,
+                "image" => $filename,
+                "description" => $request->description,
+                "owner_id" => $request->owner_id,
+            ]);
+            return response()->json([
+                "status" => true,
+                "message" => "Berhasil mengubah data gerai",
+                "body" => $outlet->getData(),
+            ], 200);
         }
-        $outlet->update([
-            "name" => $request->name,
-            "image" => $filename,
-            "description" => $request->description,
-            "owner_id" => $request->owner_id,
-        ]);
         return response()->json([
-            "status" => true,
-            "message" => "Berhasil mengubah data gerai",
-            "body" => $outlet->getData(),
-        ], 200);
+            "status" => false,
+            "message" => "Gerai tidak ditemukan",
+            "body" => [],
+        ], 404);
     }
 
     public function destroy($id) {
